requirements.txt
asgiref==3.7.2
Django==5.0.3
django-phonenumber-field==7.3.0
phonenumbers==8.13.32
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
